# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Personal Budget Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /savings:
    get:
      operationId: getSavings
      tags: [savings]
      summary: Get all savings
      responses:
        "200":
          description: List of savings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Saving"
    post:
      operationId: createSaving
      tags: [savings]
      summary: Create saving
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSavingInput"
      responses:
        "201":
          description: Created saving
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Saving"
  /savings/{id}:
    put:
      operationId: updateSaving
      tags: [savings]
      summary: Update saving
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSavingInput"
      responses:
        "200":
          description: Updated saving
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Saving"
    delete:
      operationId: deleteSaving
      tags: [savings]
      summary: Delete saving
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Notes
...
  /reminders:
    get:
      operationId: getReminders
      tags: [reminders]
      summary: Get all reminders
      responses:
        "200":
          description: List of reminders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reminder"
    post:
      operationId: createReminder
      tags: [reminders]
      summary: Create reminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReminderInput"
      responses:
        "201":
          description: Created reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
  /reminders/{id}:
    put:
      operationId: updateReminder
      tags: [reminders]
      summary: Update reminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReminderInput"
      responses:
        "200":
          description: Updated reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
    delete:
      operationId: deleteReminder
      tags: [reminders]
      summary: Delete reminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /reminders/{id}/toggle:
    patch:
      operationId: toggleReminder
      tags: [reminders]
      summary: Toggle reminder completion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"

  # Settings
...
components:
  schemas:
...
    Saving:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        type:
          type: string
          enum: [savings, asset, investment]
        targetAmount:
          type: number
          nullable: true
        currentAmount:
          type: number
        notes:
          type: string
        createdAt:
          type: string
      required: [id, name, category, type, currentAmount, notes, createdAt]
    CreateSavingInput:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        type:
          type: string
          enum: [savings, asset, investment]
        targetAmount:
          type: number
          nullable: true
        currentAmount:
          type: number
        notes:
          type: string
      required: [name, category, type, currentAmount, notes]
...
    Reminder:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        dueDate:
          type: string
          nullable: true
        isCompleted:
          type: boolean
        priority:
          type: string
          enum: [low, medium, high]
        createdAt:
          type: string
      required: [id, title, description, isCompleted, priority, createdAt]
    CreateReminderInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        dueDate:
          type: string
          nullable: true
        isCompleted:
          type: boolean
        priority:
          type: string
          enum: [low, medium, high]
      required: [title, description, isCompleted, priority]
...