# lib/api-zod/src/generated/types/createSavingInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Budget Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateSavingInputType } from "./createSavingInputType";

export interface CreateSavingInput {
  name: string;
  category: string;
  type: CreateSavingInputType;
  targetAmount?: number | null;
  currentAmount: number;
  notes: string;
}