# lib/api-zod/src/generated/types/reminder.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Budget Management API
 * OpenAPI spec version: 0.1.0
 */
import type { ReminderPriority } from "./reminderPriority";

export interface Reminder {
  id: number;
  title: string;
  description: string;
  dueDate?: string | null;
  isCompleted: boolean;
  priority: ReminderPriority;
  createdAt: string;
}