# lib/api-zod/src/generated/types/saving.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Budget Management API
 * OpenAPI spec version: 0.1.0
 */
import type { SavingType } from "./savingType";

export interface Saving {
  id: number;
  name: string;
  category: string;
  type: SavingType;
  targetAmount?: number | null;
  currentAmount: number;
  notes: string;
  createdAt: string;
}